isValid 메서드 utils/validation으로 이동 store에 있던 메서드를 utils로 옮김
